refactor: #49 extractMenu 메서드 dto로 이동
