+++ src/main/kotlin/com/asap/asapbackend/client/openapi/menu/dto/MenuOpenApiResponse.kt
@@ -1,6 +1,8 @@
 package com.asap.asapbackend.client.openapi.menu.dto
 
 import com.asap.asapbackend.batch.menu.MenuInfoProvider
+import com.asap.asapbackend.domain.menu.domain.model.Allergy
+import com.asap.asapbackend.domain.menu.domain.model.Food
 import com.asap.asapbackend.domain.school.domain.model.School
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
@@ -33,8 +35,38 @@ data class Row(
     fun toMenuInfo(school: School): MenuInfoProvider.MenuResponse {
         return MenuInfoProvider.MenuResponse(
             school = school,
-            menu = DDISH_NM,
+            menu = convertMenu(DDISH_NM),
             day = LocalDate.parse(MLSV_YMD, DateTimeFormatter.BASIC_ISO_DATE)
         )
+    }
+
+    private fun convertMenu(menus: String): List<Food> {
+        val foodList = mutableListOf<Food>()
+
+        val menuItems = menus.split("<br/>")
+        menuItems.forEach { menuItem ->
+            val nameAndAllergies = menuItem.split(" ")
+
+            val foodName = menuItem.removeSuffix(" " + nameAndAllergies.last())
+            val allergiesStr = nameAndAllergies.last().removePrefix("(").removeSuffix(")")
+
+            val allergyNumbers: List<Int> = if (allergiesStr.isNotEmpty()) {
+                if (!allergiesStr.contains(".")) {
+                    listOf(allergiesStr.toInt())
+                } else {
+                    allergiesStr.split(".").map { it.toInt() }
+                }
+            } else {
+                emptyList()
+            }
+
+            val allergies = allergyNumbers.mapNotNull { number ->
+                Allergy.entries.find { it.number == number }
+            }
+
+            val food = Food(foodName, allergies)
+            foodList.add(food)
+        }
+        return foodList
     }
 }
